move client/testutil -> x/auth/testutil
We should never break the
top-level-packages-do-not-depend-on-x-packages rule.
